requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
